+++ zellij-client/src/lib.rs
@@ -118,7 +118,6 @@ pub fn start_client(
         palette,
     };
 
-    #[cfg(not(any(feature = "test", test)))]
     let first_msg = match info {
         ClientInfo::Attach(name, force, config_options) => {
             SESSION_NAME.set(name).unwrap();
@@ -140,16 +139,6 @@ pub fn start_client(
             )
         }
     };
-    #[cfg(any(feature = "test", test))]
-    let first_msg = {
-        let _ = SESSION_NAME.set("".into());
-        ClientToServerMsg::NewClient(
-            client_attributes,
-            Box::new(opts),
-            Box::new(config_options.clone()),
-            layout,
-        )
-    };
 
     os_input.connect_to_server(&*ZELLIJ_IPC_PIPE);
     os_input.send_to_server(first_msg);
@@ -167,7 +156,6 @@ pub fn start_client(
     > = channels::bounded(50);
     let send_client_instructions = SenderWithContext::new(send_client_instructions);
 
-    #[cfg(not(any(feature = "test", test)))]
     std::panic::set_hook({
         use zellij_utils::errors::handle_panic;
         let send_client_instructions = send_client_instructions.clone();
@@ -187,6 +175,7 @@ pub fn start_client(
                 input_loop(
                     os_input,
                     config,
+                    config_options,
                     command_is_executing,
                     send_client_instructions,
                     default_mode,
@@ -242,6 +231,7 @@ pub fn start_client(
         os_input.unset_raw_mode(0);
         let goto_start_of_last_line = format!("\u{1b}[{};{}H", full_screen_ws.rows, 1);
         let restore_snapshot = "\u{1b}[?1049l";
+        os_input.disable_mouse();
         let error = format!(
             "{}\n{}{}",
             goto_start_of_last_line, restore_snapshot, backtrace
@@ -300,6 +290,7 @@ pub fn start_client(
         goto_start_of_last_line, restore_snapshot, reset_style, show_cursor, exit_msg
     );
 
+    os_input.disable_mouse();
     os_input.unset_raw_mode(0);
     let mut stdout = os_input.get_stdout_writer();
     let _ = stdout.write(goodbye_message.as_bytes()).unwrap();