+++ zellij-utils/src/input/command.rs
@@ -1,0 +1,37 @@
+//! Trigger a command
+use super::actions::Direction;
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+#[derive(Debug, Clone)]
+pub enum TerminalAction {
+    OpenFile(PathBuf),
+    RunCommand(RunCommand),
+}
+
+#[derive(Clone, Debug, Deserialize, Default, Serialize, PartialEq, Eq)]
+pub struct RunCommand {
+    pub command: PathBuf,
+    #[serde(default)]
+    pub args: Vec<String>,
+}
+
+/// Intermediate representation
+#[derive(Clone, Debug, Deserialize, Default, Serialize, PartialEq, Eq)]
+pub struct RunCommandAction {
+    #[serde(rename = "cmd")]
+    pub command: PathBuf,
+    #[serde(default)]
+    pub args: Vec<String>,
+    #[serde(default)]
+    pub direction: Option<Direction>,
+}
+
+impl From<RunCommandAction> for RunCommand {
+    fn from(action: RunCommandAction) -> Self {
+        RunCommand {
+            command: action.command,
+            args: action.args,
+        }
+    }
+}