+++ zellij-utils/src/input/options.rs
@@ -20,10 +20,16 @@ pub struct Options {
     /// Set the default mode
     #[structopt(long)]
     pub default_mode: Option<InputMode>,
+    /// Set the default shell
+    #[structopt(long, parse(from_os_str))]
+    pub default_shell: Option<PathBuf>,
     /// Set the layout_dir, defaults to
     /// subdirectory of config dir
     #[structopt(long, parse(from_os_str))]
     pub layout_dir: Option<PathBuf>,
+    #[structopt(long)]
+    #[serde(default)]
+    pub disable_mouse_mode: bool,
 }
 
 impl Options {
@@ -50,6 +56,11 @@ impl Options {
             other => other,
         };
 
+        let default_shell = match other.default_shell {
+            None => self.default_shell.clone(),
+            other => other,
+        };
+
         let layout_dir = match other.layout_dir {
             None => self.layout_dir.clone(),
             other => other,
@@ -60,11 +71,19 @@ impl Options {
             other => other,
         };
 
+        let disable_mouse_mode = if other.disable_mouse_mode {
+            true
+        } else {
+            self.disable_mouse_mode
+        };
+
         Options {
             simplified_ui,
             theme,
             default_mode,
+            default_shell,
             layout_dir,
+            disable_mouse_mode,
         }
     }